pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "value-finder"
version = "0.1.0"
description = "Deterministic UK horse racing value finder"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "sleman606"}]
dependencies = [
  "httpx>=0.27.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
include = ["value_finder*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
